Simplify the hint box code, add a shared hook
