+++ web/packages/teleport/src/Discover/Database/DownloadScript/DownloadScript.tsx
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import React, { Suspense, useEffect, useState } from 'react';
-import { Box, Flex, Text } from 'design';
+import React, { Suspense, useState } from 'react';
+import { Box, Text } from 'design';
 import * as Icons from 'design/Icon';
 import Validation, { useRule, Validator } from 'shared/components/Validation';
 
@@ -44,11 +44,10 @@ import {
   Mark,
   ResourceKind,
   TextIcon,
+  useShowHint,
 } from '../../Shared';
 
 import type { AgentStepProps } from '../../types';
-
-const SHOW_HINT_TIMEOUT = 1000 * 60 * 5; // 5 minutes
 
 export default function Container(props: AgentStepProps) {
   const hasDbLabels = props.agentMeta?.agentMatcherLabels?.length;
@@ -127,15 +126,7 @@ export function DownloadScript(
     props.agentMeta.resourceName
   );
 
-  const [showHint, setShowHint] = useState(false);
-
-  useEffect(() => {
-    if (active) {
-      const id = window.setTimeout(() => setShowHint(true), SHOW_HINT_TIMEOUT);
-
-      return () => window.clearTimeout(id);
-    }
-  }, [active]);
+  const showHint = useShowHint(active);
 
   function handleNextStep() {
     props.updateAgentMeta({
@@ -149,21 +140,7 @@ export function DownloadScript(
   let hint;
   if (showHint) {
     hint = (
-      <HintBox>
-        <Text color="warning">
-          <Flex alignItems="center" mb={2}>
-            <TextIcon
-              color="warning"
-              css={`
-                white-space: pre;
-              `}
-            >
-              <Icons.Warning fontSize={4} color="warning" />
-            </TextIcon>
-            We're still looking for your database service
-          </Flex>
-        </Text>
-
+      <HintBox header="We're still looking for your database service">
         <Text mb={3}>
           There are a couple of possible reasons for why we haven't been able to
           detect your database service.
