+++ web/packages/teleport/src/Discover/Kubernetes/KubeWrapper.tsx
@@ -11,7 +11,7 @@ const PING_INTERVAL = 1000 * 3; // 3 seconds
 export function KubeWrapper(props: WrapperProps) {
   useEffect(() => {
     return () => {
-      // once the user leaves the desktop setup flow, delete the existing token
+      // once the user leaves this flow, delete the existing token
       clearCachedJoinTokenResult(ResourceKind.Kubernetes);
     };
   }, []);
